pyproject.toml
[project]
name = "weaviate-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.0",
    "weaviate-client>=4.15.0",
]
